Model Scaling Functions

MacroEnergySystemsScaling.calc_coeff_multiplier — Method

calc_coeff_multiplier(coeff::Real, coeff_lb::Real, coeff_ub::Real)

Calculate the multiplier to scale the coefficient coeff to be within the bounds coeff_lb and coeff_ub.

MacroEnergySystemsScaling.calc_rhs_multiplier — Method

calc_rhs_multiplier(con_ref::ConstraintRef, rhs_lb::Real, rhs_ub::Real, coeff_lb::Real, coeff_ub::Real)

Calculate the multiplier to scale the right-hand side of the constraint con_ref to be within the bounds rhs_lb and rhs_ub.

MacroEnergySystemsScaling.existing_proxy_var — Method

existing_proxy_var(var::VariableRef, multiplier::Real, proxy_var_map::Dict{VariableRef, Vector{Tuple{VariableRef, Float64}}}, proxy_var_ratio_ub::Real)

Check if a proxy variable already exists for the variable var with a multiplier close to multiplier. If such a proxy variable exists, return it and its multiplier; otherwise, return nothing.

MacroEnergySystemsScaling.get_proxy_var — Method

get_proxy_var(var::VariableRef, multiplier::Real, proxy_var_map::Dict{VariableRef, Vector{Tuple{VariableRef, Float64}}}, proxy_var_ratio_ub::Real)

Check if a cached proxy variable exists for the variable var with a multiplier close to multiplier. If such a proxy variable exists, return it and its multiplier; otherwise, create a new proxy variable and return it.

MacroEnergySystemsScaling.make_constraint — Function

make_constraint(EP::Model, var_coeff_pairs::AbstractDict{VariableRef, Float64}, rhs::MOI.AbstractScalarSet, con_name::AbstractString, rhs_multiplier::Real=1.0)

Create a new constraint with the given variable-coefficient pairs, right-hand side, and name; then add it to the model EP.

MacroEnergySystemsScaling.make_proxy_var — Method

make_proxy_var(var::VariableRef, multiplier::Real)

Create a new proxy variable for the variable var with the given multiplier. The proxy and original variable are related by: var == proxy_var * multiplier.

MacroEnergySystemsScaling.parse_constraint_name — Method

parse_constraint_name(input::AbstractString)

Parse the name of a constraint and return the name and indexes as a tuple.

MacroEnergySystemsScaling.prune_coefficients — Method

prune_coefficients(new_coeff::Real, coeff::Real, multiplier::Real)

If a new coefficient is close to 1.0 or -1.0, return 1.0 or -1.0 respectively. "close" is defined using the Julia isapprox function.

MacroEnergySystemsScaling.replace_constraint! — Function

replace_constraint!(con_ref::ConstraintRef, var_coeff_pairs=nothing, rhs_multiplier::Real=1.0)

Replace the constraint con_ref with a new constraint with the given variable-coefficient pairs and right-hand side multiplier.

MacroEnergySystemsScaling.scale_and_remake_constraint — Method

scale_and_remake_constraint(con_ref::ConstraintRef, scaling_settings::ScalingSettings)

Scale the coefficients and RHS of the constraint con_ref using the scaling settings scaling_settings. First we check if we can scale the right-hand side constant without creating proxy variables. Afterwards, we iterate over the variable-coefficient pairs and scale them using proxy variables if necessary. The original constraint is then replaced with a new constraint using the scaled variable-coefficient pairs.

MacroEnergySystemsScaling.scale_constraint! — Method

scale_constraint!(con_ref::ConstraintRef, scaling_settings::ScalingSettings)

Scale the coefficients and RHS of the constraint con_ref using the scaling settings scaling_settings. con_ref is a JuMP constraint reference.

MacroEnergySystemsScaling.scale_constraints! — Function

scale_constraints!(EP::Model, scaling_settings::ScalingSettings=ScalingSettings())

Scale the coefficients and RHS of all constraints in the model EP using the scaling settings scaling_settings. This function creates an array of all constraints in EP and then broadcasts scaleconstraint!(conref, scaling_settings) on the array.

MacroEnergySystemsScaling.scale_constraints! — Function

scale_constraints!(constraint_list::Vector{ConstraintRef}, scaling_settings::ScalingSettings=ScalingSettings())

Scale the coefficients and RHS of all constraints in the model EP using the scaling settings scaling_settings. This function calls scaleconstraint!(conref, scalingsettings) on each constraint in `constraintlist`.

MacroEnergySystemsScaling.update_var_coeff_pair — Method

update_var_coeff_pair(var::VariableRef, coeff::Real, scaling_settings::ScalingSettings)

Update the variable-coefficient pair (var, coeff) to be within the bounds specified in scaling_settings.